Rename GroupingServicePluging to GroupingPlugin
The old name was confusing as we already have a grouping_service.

The main focus of the pluging is not to abstract the existing service
(although it might use it) but to abstract the API of the different LMS.
