+++ lms/product/plugin/__init__.py
@@ -1,8 +1,8 @@
-from lms.product.plugin.grouping_service import GroupingServicePlugin
+from lms.product.plugin.grouping import GroupingPlugin
 from lms.product.plugin.plugin import PluginConfig, Plugins
 
 
 def includeme(config):  # pragma: nocover
     """Register all of our plugins."""
 
-    config.register_service(GroupingServicePlugin(), iface=GroupingServicePlugin)
+    config.register_service(GroupingPlugin(), iface=GroupingPlugin)
